+++ CHANGELOG.md
@@ -139,3 +139,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Update the dialog stepper only if necessary
 - consider only selected forms for mandatory fields in the project view
+
+### Changed
+- Notification only if project manager admin